1) mengubah db; 2) menambah kelola tipe kamar di admin; 3) tambah kolom status_kmr pd tabel kamar; 4) mengubah invoice yg baru; 5) tipe kamar berelasi dengan fasilitas kamar

+++ app/Views/petugas/fasilitas-umum.php
@@ -31,6 +31,12 @@
                     <strong><?= session()->getFlashdata('edit_fumum'); ?></strong>
                 </div>
             <?php endif; ?>
+            <?php if (session()->getFlashdata('hapus_fumum')) : ?>
+                <div class="alert alert-success alert-dismissible fade show" role="alert">
+                    <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+                    <strong><?= session()->getFlashdata('hapus_fumum'); ?></strong>
+                </div>
+            <?php endif; ?>
             <?php if (!$data_fumum) { ?>
                 <div class="alert alert-danger" role="alert">
                     <strong>Data fasilitas hotel tidak ditemukan!</strong>
@@ -41,7 +47,7 @@
                     <thead>
                         <tr class="text-center">
                             <th>#</th>
-                            <th>Fasilitas Hotel</th>
+                            <th width="200px">Fasilitas Hotel</th>
                             <th>Foto</th>
                             <th>Deskripsi</th>
                             <th>Aksi</th>
@@ -52,14 +58,14 @@
                         foreach ($data_fumum as $row) : ?>
                             <tr>
                                 <td class="text-center"><?= $no++; ?></td>
-                                <td><?= $row['nama_fumum']; ?></td>
+                                <td class="text-center"><?= $row['nama_fumum']; ?></td>
                                 <td>
                                     <img width="300px" src="/gambar/<?= $row['foto']; ?>" alt="">
                                 </td>
                                 <td><?= $row['deskripsi']; ?></td>
                                 <td class="text-center">
                                     <a href="/petugas/fumum/edit/<?= $row['id_fumum']; ?>" class="btn btn-warning btn-sm mx-1 my-1">Edit</a>
-                                    <a href="/petugas/fumum/hapus/<?= $row['id_fumum']; ?>" class="btn btn-danger  btn-sm mx-1 my-1">Hapus</a>
+                                    <a href="/petugas/fumum/hapus/<?= $row['id_fumum']; ?>" onclick="return confirm('Apakah Anda yakin akan menghapusnya?')" class="btn btn-danger  btn-sm mx-1 my-1">Hapus</a>
                                 </td>
                             </tr>
                         <?php endforeach; ?>
